add tests

+++ Src/libCZI/CreateBitmap.cpp
@@ -31,52 +31,52 @@ using namespace libCZI;
 
 static std::shared_ptr<libCZI::IBitmapData> CreateBitmapFromSubBlock_JpgXr(ISubBlock* subBlk)
 {
-	auto dec = GetSite()->GetDecoder(ImageDecoderType::JPXR_JxrLib, nullptr);
-	const void* ptr; size_t size;
-	subBlk->DangerousGetRawData(ISubBlock::MemBlkType::Data, ptr, size);
-	return dec->Decode(ptr, size);
+    auto dec = GetSite()->GetDecoder(ImageDecoderType::JPXR_JxrLib, nullptr);
+    const void* ptr; size_t size;
+    subBlk->DangerousGetRawData(ISubBlock::MemBlkType::Data, ptr, size);
+    return dec->Decode(ptr, size);
 }
 
 static std::shared_ptr<libCZI::IBitmapData> CreateBitmapFromSubBlock_Uncompressed(ISubBlock* subBlk)
 {
-	size_t size;
-	CSharedPtrAllocator sharedPtrAllocator(subBlk->GetRawData(ISubBlock::MemBlkType::Data, &size));
+    size_t size;
+    CSharedPtrAllocator sharedPtrAllocator(subBlk->GetRawData(ISubBlock::MemBlkType::Data, &size));
 
-	const auto& sbBlkInfo = subBlk->GetSubBlockInfo();
+    const auto& sbBlkInfo = subBlk->GetSubBlockInfo();
 
-	// TODO: How exactly shoud the stride be derived? It seems that stride must be exactly linesize.
-	const std::uint32_t stride = sbBlkInfo.physicalSize.w * CziUtils::GetBytesPerPel(sbBlkInfo.pixelType);
-	if (stride * sbBlkInfo.physicalSize.h > size)
-	{
-		throw std::logic_error("insufficient size of subblock");
-	}
+    // TODO: How exactly shoud the stride be derived? It seems that stride must be exactly linesize.
+    const std::uint32_t stride = sbBlkInfo.physicalSize.w * CziUtils::GetBytesPerPel(sbBlkInfo.pixelType);
+    if ((size_t)stride * sbBlkInfo.physicalSize.h > size)
+    {
+        throw std::logic_error("insufficient size of subblock");
+    }
 
-	auto sb = CBitmapData<CSharedPtrAllocator>::Create(
-		sharedPtrAllocator,
-		sbBlkInfo.pixelType,
-		sbBlkInfo.physicalSize.w,
-		sbBlkInfo.physicalSize.h,
-		stride);
+    auto sb = CBitmapData<CSharedPtrAllocator>::Create(
+        sharedPtrAllocator,
+        sbBlkInfo.pixelType,
+        sbBlkInfo.physicalSize.w,
+        sbBlkInfo.physicalSize.h,
+        stride);
 
 #if LIBCZI_ISBIGENDIANHOST
-	if (!CziUtils::IsPixelTypeEndianessAgnostic(subBlk->GetSubBlockInfo().pixelType))
-	{
-		return CBitmapOperations::ConvertToBigEndian(sb.get());
-	}
+    if (!CziUtils::IsPixelTypeEndianessAgnostic(subBlk->GetSubBlockInfo().pixelType))
+    {
+        return CBitmapOperations::ConvertToBigEndian(sb.get());
+    }
 #endif
 
-	return sb;
+    return sb;
 }
 
 std::shared_ptr<libCZI::IBitmapData> libCZI::CreateBitmapFromSubBlock(ISubBlock* subBlk)
 {
-	switch (subBlk->GetSubBlockInfo().mode)
-	{
-	case CompressionMode::JpgXr:
-		return CreateBitmapFromSubBlock_JpgXr(subBlk);
-	case CompressionMode::UnCompressed:
-		return CreateBitmapFromSubBlock_Uncompressed(subBlk);
-	default:	// silence warnings
-		throw std::logic_error("The method or operation is not implemented.");
-	}
+    switch (subBlk->GetSubBlockInfo().mode)
+    {
+    case CompressionMode::JpgXr:
+        return CreateBitmapFromSubBlock_JpgXr(subBlk);
+    case CompressionMode::UnCompressed:
+        return CreateBitmapFromSubBlock_Uncompressed(subBlk);
+    default:	// silence warnings
+        throw std::logic_error("The method or operation is not implemented.");
+    }
 }
